Updated page objects
