+++ automated_tests_selenium/page_objects/childcare_address_task.py
@@ -1,19 +1,16 @@
-from django.test import LiveServerTestCase
+from django.test import TestCase
 
 
-class ChildcareAddressTask(LiveServerTestCase):
+class ChildcareAddressTask(TestCase):
 
-
-    def __init__(self, task_executor):
+    def __init__(self, task_executor, *args, **kwargs):
         """
         Default constructor
-        :param driver: the selenium driver to be used for executing steps
-        :param base_url: the base url against which tests are to be executed
         """
-
         self.task_executor = task_executor
+        super(ChildcareAddressTask, self).__init__(*args, **kwargs)
 
-    def same_as_applicant_address(self,):
+    def complete_childcare_address(self):
         driver = self.task_executor.get_driver()
         self.assertEqual("To do", driver.find_element_by_xpath("//tr[@id='childcare_address']/td/a/strong").text)
         self.task_executor.click_element_by_xpath("//tr[@id='childcare_address']/td/a/span")
@@ -25,4 +22,3 @@ class ChildcareAddressTask(LiveServerTestCase):
         self.task_executor.click_element_by_xpath("//input[@value='Save and continue']")
         self.task_executor.click_element_by_xpath("//*[@value='Confirm and continue']")
         self.assertEqual("Done", driver.find_element_by_xpath("//tr[@id='childcare_address']/td/a/strong").text)
-
